+++ source/buttons/secondtoast.js
@@ -135,7 +135,14 @@ module.exports = new ButtonWrapper(mainId, 3000,
 					thread.send({ content, flags: MessageFlags.SuppressNotifications });
 				})
 			}
-			company.updateScoreboard(interaction.guild, logicLayer);
+			const embeds = [];
+			const ranks = await logicLayer.ranks.findAllRanks(interaction.guild.id);
+			if (company.scoreboardIsSeasonal) {
+				embeds.push(await company.seasonalScoreboardEmbed(interaction.guild, await logicLayer.seasons.findSeasonParticipations(season.id), ranks));
+			} else {
+				embeds.push(await company.overallScoreboardEmbed(interaction.guild, await logicLayer.hunters.findCompanyHunters(interaction.guild.id), ranks));
+			}
+			company.updateScoreboard(interaction.guild, embeds);
 		})
 
 		if (progressData.goalCompleted) {