+++ source/commands/bounty/post.js
@@ -176,9 +176,16 @@ module.exports = new SubcommandWrapper("post", "Post your own bounty (+1 XP)",
 				logicLayer.seasons.changeSeasonXP(modalSubmission.user.id, modalSubmission.guildId, season.id, 1);
 				const company = await logicLayer.companies.findCompanyByPK(modalSubmission.guild.id);
 				const poster = await logicLayer.hunters.findOneHunter(modalSubmission.user.id, modalSubmission.guildId);
-				poster.increment({ xp: 1 }).then(() => {
+				poster.increment({ xp: 1 }).then(async () => {
 					getRankUpdates(modalSubmission.guild, logicLayer);
-					company.updateScoreboard(interaction.guild, logicLayer);
+					const embeds = [];
+					const ranks = await logicLayer.ranks.findAllRanks(interaction.guild.id);
+					if (company.scoreboardIsSeasonal) {
+						embeds.push(await company.seasonalScoreboardEmbed(interaction.guild, await logicLayer.seasons.findSeasonParticipations(season.id), ranks));
+					} else {
+						embeds.push(await company.overallScoreboardEmbed(interaction.guild, await logicLayer.hunters.findCompanyHunters(interaction.guild.id), ranks));
+					}
+					company.updateScoreboard(interaction.guild, embeds);
 				});
 
 				if (shouldMakeEvent) {